+++ src/base_object.h
@@ -31,6 +31,7 @@
 namespace node {
 
 class Environment;
+class IsolateData;
 template <typename T, bool kIsWeak>
 class BaseObjectPtrImpl;
 
@@ -111,8 +112,10 @@ class BaseObject : public MemoryRetainer {
   // a BaseObjectPtr to this object.
   inline void Detach();
 
-  static v8::Local<v8::FunctionTemplate> GetConstructorTemplate(
+  static inline v8::Local<v8::FunctionTemplate> GetConstructorTemplate(
       Environment* env);
+  static v8::Local<v8::FunctionTemplate> GetConstructorTemplate(
+      IsolateData* isolate_data);
 
   // Interface for transferring BaseObject instances using the .postMessage()
   // method of MessagePorts (and, by extension, Workers).
