+++ src/node_context_data.h
@@ -36,11 +36,15 @@ namespace node {
 #define NODE_CONTEXT_CONTEXTIFY_CONTEXT_INDEX 37
 #endif
 
+#ifndef NODE_CONTEXT_REALM_INDEX
+#define NODE_CONTEXT_REALM_INDEX 38
+#endif
+
 // NODE_CONTEXT_TAG must be greater than any embedder indexes so that a single
 // check on the number of embedder data fields can assure the presence of all
 // embedder indexes.
 #ifndef NODE_CONTEXT_TAG
-#define NODE_CONTEXT_TAG 38
+#define NODE_CONTEXT_TAG 39
 #endif
 
 enum ContextEmbedderIndex {
@@ -51,6 +55,7 @@ enum ContextEmbedderIndex {
   kAllowCodeGenerationFromStrings =
       NODE_CONTEXT_ALLOW_CODE_GENERATION_FROM_STRINGS_INDEX,
   kContextifyContext = NODE_CONTEXT_CONTEXTIFY_CONTEXT_INDEX,
+  kRealm = NODE_CONTEXT_REALM_INDEX,
   kContextTag = NODE_CONTEXT_TAG,
 };
