+++ test/pummel/test-heapdump-env.js
@@ -16,8 +16,14 @@ const context = require('vm').createScript('const foo = 123');
 validateSnapshotNodes('Node / Environment', [{
   children: [
     { node_name: 'Node / cleanup_hooks', edge_name: 'cleanup_hooks' },
-    { node_name: 'process', edge_name: 'process_object' },
     { node_name: 'Node / IsolateData', edge_name: 'isolate_data' },
+    { node_name: 'Node / Realm', edge_name: 'principal_realm' },
+  ]
+}]);
+
+validateSnapshotNodes('Node / Realm', [{
+  children: [
+    { node_name: 'process', edge_name: 'process_object' },
   ]
 }]);
